+++ enemies/zombie.gd
@@ -154,14 +154,17 @@ func force_state_for_test(next_state: ZombieState) -> void:
 	_change_state(next_state)
 
 func _configure_nodes() -> void:
+	map_collision_mask &= ~PLAYER_BOUNDARY_BLOCKER_COLLISION_LAYER
+	vision_collision_mask &= ~PLAYER_BOUNDARY_BLOCKER_COLLISION_LAYER
 	if zombie_body != null:
 		zombie_body.collision_layer |= ZOMBIE_COLLISION_LAYER
-		zombie_body.collision_mask |= (
-			WORLD_COLLISION_LAYER
+		zombie_body.collision_mask = (
+			zombie_body.collision_mask
+			| WORLD_COLLISION_LAYER
 			| PLAYER_COLLISION_LAYER
 			| ZOMBIE_COLLISION_LAYER
 			| SKELETON_COLLISION_LAYER
-		)
+		) & ~PLAYER_BOUNDARY_BLOCKER_COLLISION_LAYER
 
 	if navigation_agent != null:
 		navigation_agent.path_desired_distance = 0.18